+++ plugin/src/main/java/de/infolektuell/gradle/jextract/extensions/JextractExtension.java
@@ -1,25 +1,17 @@
 package de.infolektuell.gradle.jextract.extensions;
 
-import org.gradle.api.Action;
-import org.gradle.api.NamedDomainObjectContainer;
 import org.gradle.api.file.Directory;
 import org.gradle.api.file.DirectoryProperty;
 import org.gradle.api.file.RegularFileProperty;
-import org.gradle.api.model.ObjectFactory;
-import org.gradle.api.provider.Property;
 import org.gradle.api.tasks.Nested;
 import org.gradle.jvm.toolchain.JavaLanguageVersion;
 import org.jspecify.annotations.NonNull;
 
-import javax.inject.Inject;
-
+/**
+ * Configures the Jextract installation location.
+ */
 public abstract class JextractExtension {
     public static final String EXTENSION_NAME = "jextract";
-    private final NamedDomainObjectContainer<@NonNull LibraryHandler> libraries;
-    public JextractExtension() {
-        super();
-        this.libraries = getObjects().domainObjectContainer(LibraryHandler.class);
-    }
 
     /**
      * Configuration of a Jextract installation to be used by the plugin
@@ -52,30 +44,4 @@ public abstract class JextractExtension {
      * A [properties][java.util.Properties] file containing the remote locations where to download the Jextract distributions
      */
     public abstract RegularFileProperty getDistributions();
-
-    /**
-     * The libraries Jextract should generate bindings for
-     */
-    public final NamedDomainObjectContainer<@NonNull LibraryHandler> getLibraries() {
-        return this.libraries;
-    }
-
-    /**
-     * Configures the libraries Jextract should generate bindings for
-     */
-    public final void libraries(@NonNull Action<@NonNull NamedDomainObjectContainer<@NonNull LibraryHandler>> action) {
-        action.execute(this.libraries);
-    }
-
-    /**
-     * specify the directory to place generated files of all libraries by default
-     */
-    public abstract DirectoryProperty getOutput();
-
-    /**
-     * Generate source files instead of class files for all libraries where not set explicitly (Jextract 21 and below)
-     */
-    public abstract Property<@NonNull Boolean> getGenerateSourceFiles();
-
-    @Inject protected abstract ObjectFactory getObjects();
 }